dpif: add 'dpif-bpf' provider.
Implement a new datapath interface for use with BPF datapaths.

Like dpif-netlink, dpif-bpf is backed by an implementation which resides
within the kernel. It uses the BPF functionality available in recent
versions of Linux to create the datapath. Unlike dpif-netlink there is no
datapath notion of a bridge with ports attached; dpif-bpf is implemented
by attaching BPF programs directly to individual devices using TC.

Upcalls are implemented using a perf event ringbuffer, which is polled
by handler threads. Flow execution is implemented by sending the packet
plus metadata on a dedicated tap device, where there is an BPF program
that understands the format of the packet coming from userspace. When
this device receives a message, it strips the metadata, uses it to
determine how to execute the packet, then forwards the packet onwards.

This initial implementation has a number of limitations which are
expected to go away over time:
	* The set of matches and actions supported by the datapath is not
	  as wide as the full set known by OVS, so if a flow cannot be
	  expressed in the current eBPF API, OVS will log errors and return
	  errors during flow put.
	* Only the input port and packet length is passed as metadata from
	  the datapath to userspace during upcall. Key extraction is done
	  purely from the packet provided from the datapath.
	* Conversely, only the output port is sent down during execution.
	  No other actions are supported currently; and only one output is
	  supported.
	* Ingress policing cannot be configured on BPF
	  datapath devices.
	* On startup, if the OVS BPF datapath is already loaded into the
	  kernel and pinned to the filesystem, it will reuse this datapath,
	  even if the datapath is out-of-date.

Documentation/intro/install/bpf.rst contains further information on how
to build and use the bpf datapath.

For more details on the design and implementation, see our OSR paper:
[1] https://dl.acm.org/citation.cfm?id=3139657
[2] http://openvswitch.org/support/ovscon2016/7/1120-tu.pdf

Signed-off-by: Joe Stringer <joe@ovn.org>
Signed-off-by: William Tu <u9012063@gmail.com>
Signed-off-by: Yifeng Sun <pkusunyifeng@gmail.com>
Co-authored-by: William Tu <u9012063@gmail.com>
Co-authored-by: Yifeng Sun <pkusunyifeng@gmail.com>