bpf: implement OVS BPF datapath.
This patch adds the OVS-eBPF datapath implementation for dpif-bpf.
Three stages are added: parse, lookup, and actions.  Each stages are
tail called to the next stage.  When executing multiple actions,
the current action also tail calls the subsequent action, based on
the result of flow table lookup.

The protocol headers are auto-generated and defined at generated_headers.h.
The bpf_flow_key is extracted using the P4-to-eBPF compiler from
the bcc project.  A couple of manual tweaks are required, see parser.h.

Signed-off-by: William Tu <u9012063@gmail.com>
Signed-off-by: Yifeng Sun <pkusunyifeng@gmail.com>
Signed-off-by: Joe Stringer <joe@ovn.org>
Co-authored-by: Joe Stringer <joe@ovn.org>
Co-authored-by: Yifeng Sun <pkusunyifeng@gmail.com>