bpf: Support megaflow
This patch provides basic megaflow support for BPF datapath.

Two BPF maps are introduced for this implementation.
  * megaflow_mask_table is an array that stores a list of flow masks.
  * megaflow_table is a hash map that maps masked flow key to the
    associated action batch.

With this change, in the BPF datapath, we first look up the exact match
cache (EMC) flow table.  If we miss the EMC, we then look up the megaflow
table for flow match.  While looking up the megaflow table, we iterate
through the list of megaflow mask, apply the mask to flow key, and look up
megaflow table with the masked flow key.  We execute the action batch if
there is a hit, otherwise the packet is send to ovs-vswitchd for flow
lookup through the upcall mechanism via perf event ring buffer.

Some changes are made in dpif-bpf-odp to accommodate bpf flow mask
generation.  Some modification also take place in dpif-bpf to take
care of megaflow revalidation.

Limitation
  * Currently, the maximum number of megaflow mask is 28.

Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>